box: change schema version type to `uint64_t`
The requirement of the schema version is that it monotonically grows: in
order to protect from wrapping, change its type and the type of everything
depending on it to `uint64_t`.

Needed for #7904

NO_CHANGELOG=enhancement
NO_DOC=enhancement
NO_TEST=enhancement

+++ src/box/iproto.cc
@@ -1793,7 +1793,7 @@ net_finish_destroy(struct cmsg *m)
 
 
 static int
-tx_check_schema(uint32_t new_schema_version)
+tx_check_schema(uint64_t new_schema_version)
 {
 	if (new_schema_version && new_schema_version != schema_version) {
 		diag_set(ClientError, ER_WRONG_SCHEMA_VERSION,
