+++ test/unit/vy_point_lookup.c
@@ -10,7 +10,7 @@
 #include "vy_write_iterator.h"
 #include "identifier.h"
 
-uint32_t schema_version;
+uint64_t schema_version;
 uint32_t space_cache_version;
 struct space *space_by_id(uint32_t id) { return NULL; }
 struct vy_lsm *vy_lsm(struct index *index) { return NULL; }
